# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MarvelousBlog-Backend Swagger接口文档",
        "title": "PP同学接口文档",
        "contact": {
            "name": "PP同学",
            "email": "710955321@qq.com"
        },
        "license": {},
        "version": "0.1"
    },
    "paths": {
        "/v1/article/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建一篇新的文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article接口"
                ],
                "summary": "新增文章",
                "parameters": [
                    {
                        "description": "文章内容",
                        "name": "label",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Article"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "新增文章成功",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    }
                }
            }
        },
        "/v1/article/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "输入文章id以删除文章（软删除，并未在数据库中实际删除）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article接口"
                ],
                "summary": "删除某篇文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "删除的article的id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "删除成功"
                    },
                    "400": {
                        "description": "输入参数有误",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "402": {
                        "description": "用户未授权",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    }
                }
            }
        },
        "/v1/article/detail/{id}": {
            "get": {
                "description": "获取某篇文章的内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article接口"
                ],
                "summary": "获取某篇文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "所请求的文章id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Article"
                            }
                        }
                    },
                    "400": {
                        "description": "输入参数有误",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "404": {
                        "description": "未找到资源",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    }
                }
            }
        },
        "/v1/article/enable/{id}": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "管理员重新启用某篇文章（status 0 -\u003e 1）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article接口"
                ],
                "summary": "管理员Enable文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "enable的文章id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Enable成功",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "400": {
                        "description": "输入参数有误",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "404": {
                        "description": "未找到资源",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    }
                }
            }
        },
        "/v1/article/list": {
            "get": {
                "description": "获取某篇文章的内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article接口"
                ],
                "summary": "分页获取全部文章(每页10篇）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "请求的页码数",
                        "name": "pageNum",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Article"
                            }
                        }
                    },
                    "400": {
                        "description": "输入参数有误",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "404": {
                        "description": "未找到资源",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    }
                }
            }
        },
        "/v1/article/modify/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "输入新的文章内容以更新文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article接口"
                ],
                "summary": "修改文章内容",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "修改的文章的id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "修改的文章信息",
                        "name": "visitor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Article"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "402": {
                        "description": "用户未授权",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "404": {
                        "description": "未找到资源",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    }
                }
            }
        },
        "/v1/article/onLabel/{labelId}": {
            "get": {
                "description": "获取某一标签下的全部文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article接口"
                ],
                "summary": "获取某一标签下的全部文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "请求的标签的id",
                        "name": "labelId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Article"
                            }
                        }
                    },
                    "400": {
                        "description": "输入参数有误",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "404": {
                        "description": "未找到资源",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    }
                }
            }
        },
        "/v1/article/superAuthor/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "管理员获取全部文章（包括status = 0的）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Article接口"
                ],
                "summary": "管理员获取全部文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "请求的页码数",
                        "name": "pageNum",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Article"
                            }
                        }
                    },
                    "400": {
                        "description": "输入参数有误",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    },
                    "404": {
                        "description": "未找到资源",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseObject"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.ResponseObject": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "业务响应码（非HTTP码）"
                },
                "message": {
                    "type": "string",
                    "example": "Some message about the code."
                }
            }
        },
        "model.Article": {
            "type": "object",
            "properties": {
                "authorId": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "labelId": {
                    "type": "integer"
                },
                "lastModifyTime": {
                    "type": "string"
                },
                "postTime": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "summary": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "visitCount": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}